Delete now tests alongside Update
