+++ src/Subjects/Specs.Integration/Business/DeleteBusinessCommand.feature
@@ -13,7 +13,7 @@ Scenario: Delete business
 
 Examples:
 	| def								| response		| responseErrors                        | requestBusinessKey					|
-	| success TaxNumber BusinessKey		| Success		|                                       | d1604a05-f883-40f1-803b-8562b5674f1a  |
-	| not found BusinessKey				| NotFound		|										| 12345600-1234-1234-1234-123400000000	|
+	| success TaxNumber BusinessKey		| Success		|                                       | 038213ba-9d95-42f3-8e8b-126cec10481b  |
+	| not found BusinessKey				| NotFound		|										| 038213ba-9d95-42f3-8e8b-126cec10481b	|
 	| bad request Empty BusinessKey		| BadRequest	| BusinessKey							| 00000000-0000-0000-0000-000000000000	|
 	| bad request non-guid BusinessKey	| BadRequest	| BusinessKey							| 11111									|
