+++ src/Subjects/Specs.Integration/Business/GetBusinessesByName.feature.cs
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace Goodtocode.Subjects.Integration.Business.Commands
+namespace Goodtocode.Subjects.Integration.Business
 {
     using TechTalk.SpecFlow;
     using System;
@@ -20,25 +20,25 @@ namespace Goodtocode.Subjects.Integration.Business.Commands
     [System.CodeDom.Compiler.GeneratedCodeAttribute("TechTalk.SpecFlow", "3.9.0.0")]
     [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     [NUnit.Framework.TestFixtureAttribute()]
-    [NUnit.Framework.DescriptionAttribute("Delete Business Command")]
-    [NUnit.Framework.CategoryAttribute("deleteBusinessCommand")]
-    public partial class DeleteBusinessCommandFeature
+    [NUnit.Framework.DescriptionAttribute("Find Businesses by name")]
+    [NUnit.Framework.CategoryAttribute("getBusinessesByName")]
+    public partial class FindBusinessesByNameFeature
     {
         
         private TechTalk.SpecFlow.ITestRunner testRunner;
         
         private static string[] featureTags = new string[] {
-                "deleteBusinessCommand"};
+                "getBusinessesByName"};
         
-#line 1 "DeleteBusinessCommand.feature"
+#line 1 "GetBusinessesByName.feature"
 #line hidden
         
         [NUnit.Framework.OneTimeSetUpAttribute()]
         public virtual void FeatureSetup()
         {
             testRunner = TechTalk.SpecFlow.TestRunnerManager.GetTestRunner();
-            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "Business/Commands", "Delete Business Command", "As an customer service agent\r\nI delete the business\r\nThe business is deleted from" +
-                    " the system of record", ProgrammingLanguage.CSharp, featureTags);
+            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "Business", "Find Businesses by name", "\tAs a customer service rep\r\n\tWhen I search for a business by name critera\r\n\tI can" +
+                    " see a list of businesses that match the name criteria", ProgrammingLanguage.CSharp, featureTags);
             testRunner.OnFeatureStart(featureInfo);
         }
         
@@ -77,20 +77,19 @@ namespace Goodtocode.Subjects.Integration.Business.Commands
         }
         
         [NUnit.Framework.TestAttribute()]
-        [NUnit.Framework.DescriptionAttribute("Delete business")]
-        [NUnit.Framework.TestCaseAttribute("success TaxNumber BusinessKey", "Success", "", "d1604a05-f883-40f1-803b-8562b5674f1a", null)]
-        [NUnit.Framework.TestCaseAttribute("not found BusinessKey", "NotFound", "", "12345600-1234-1234-1234-123400000000", null)]
-        [NUnit.Framework.TestCaseAttribute("bad request Empty BusinessKey", "BadRequest", "BusinessKey", "00000000-0000-0000-0000-000000000000", null)]
-        [NUnit.Framework.TestCaseAttribute("bad request non-guid BusinessKey", "BadRequest", "BusinessKey", "11111", null)]
-        public void DeleteBusiness(string def, string response, string responseErrors, string requestBusinessKey, string[] exampleTags)
+        [NUnit.Framework.DescriptionAttribute("Get Existing Business By Name")]
+        [NUnit.Framework.TestCaseAttribute("success exists", "Success", "", "BusinessInDb", "true", null)]
+        [NUnit.Framework.TestCaseAttribute("empty name", "BadRequest", "BusinessName", "", "false", null)]
+        public void GetExistingBusinessByName(string def, string response, string responseErrors, string businessName, string businessExists, string[] exampleTags)
         {
             string[] tagsOfScenario = exampleTags;
             System.Collections.Specialized.OrderedDictionary argumentsOfScenario = new System.Collections.Specialized.OrderedDictionary();
             argumentsOfScenario.Add("def", def);
             argumentsOfScenario.Add("response", response);
             argumentsOfScenario.Add("responseErrors", responseErrors);
-            argumentsOfScenario.Add("requestBusinessKey", requestBusinessKey);
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Delete business", null, tagsOfScenario, argumentsOfScenario, featureTags);
+            argumentsOfScenario.Add("businessName", businessName);
+            argumentsOfScenario.Add("businessExists", businessExists);
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Get Existing Business By Name", null, tagsOfScenario, argumentsOfScenario, featureTags);
 #line 7
 this.ScenarioInitialize(scenarioInfo);
 #line hidden
@@ -105,16 +104,25 @@ this.ScenarioInitialize(scenarioInfo);
  testRunner.Given(string.Format("I have a def \"{0}\"", def), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Given ");
 #line hidden
 #line 9
- testRunner.And(string.Format("I have a BusinessKey \"{0}\"", requestBusinessKey), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
+ testRunner.And(string.Format("I have a BusinessName \"{0}\"", businessName), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
 #line 10
- testRunner.When("I delete the business", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "When ");
+ testRunner.And(string.Format("the business exists \"{0}\"", businessExists), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
 #line 11
- testRunner.Then(string.Format("The response is \"{0}\"", response), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Then ");
+ testRunner.When("I query for matching Businesses", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "When ");
 #line hidden
 #line 12
- testRunner.And(string.Format("If the response has validation issues I see the \"{0}\" in the response", responseErrors), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
+ testRunner.Then(string.Format("the response is \"{0}\"", response), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Then ");
+#line hidden
+#line 13
+ testRunner.And(string.Format("if the response has validation issues I see the \"{0}\" in the response", responseErrors), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
+#line hidden
+#line 14
+ testRunner.And("if the response is valid then the response contains a collection of businesses", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
+#line hidden
+#line 15
+ testRunner.And(string.Format("each business has a matching BusinessName of \"{0}\"", businessName), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
             }
             this.ScenarioCleanup();
