+++ src/Subjects/Specs.Unit/Business/AddBusinessCommandStepDefinitions.cs
@@ -6,15 +6,16 @@ using Moq;
 using System.Collections.Concurrent;
 using static Goodtocode.Subjects.Unit.Common.ResponseTypes;
 
-namespace Goodtocode.Subjects.Unit.Business.Commands;
+namespace Goodtocode.Subjects.Unit.Business;
 
 [Binding]
-[Scope(Tag = "deleteBusinessCommand")]
-public class DeleteBusinessCommandStepDefinitions : TestBase
+[Scope(Tag = "addBusinessCommand")]
+public class AddBusinessCommandStepDefinitions : TestBase
 {
     private IDictionary<string, string[]> _commandErrors = new ConcurrentDictionary<string, string[]>();
     private string[]? _expectedInvalidFields;
-    private Guid _businessKey;
+    private string _businessName = string.Empty;
+    private string _taxNumber = string.Empty;
     private object _responseType = string.Empty;
     private ValidationResult _validationErrors = new();
 
@@ -24,30 +25,37 @@ public class DeleteBusinessCommandStepDefinitions : TestBase
         _def = def;
     }
 
-    [Given(@"I have a BusinessKey ""([^""]*)""")]
-    public void GivenIHaveABusinessKey(string businessKey)
+    [Given(@"I have a BusinessName ""([^""]*)""")]
+    public void GivenIHaveABusinessName(string businessName)
     {
-        Guid.TryParse(businessKey, out _businessKey);
+        _businessName = businessName;
     }
 
-    [When(@"I delete the business")]
-    public async Task WhenIDeleteTheBusiness()
+    [Given(@"I have a TaxNumber ""([^""]*)""")]
+    public void GivenIHaveATaxNumber(string taxNumber)
+    {
+        _taxNumber = taxNumber;
+    }
+
+    [When(@"I add the business")]
+    public async Task WhenIAddTheBusiness()
     {
         var userBusinessRepoMock = new Mock<IBusinessRepo>();
 
-        var request = new DeleteBusinessCommand
+        var request = new AddBusinessCommand
         {
-            BusinessKey = _businessKey,
+            BusinessName = _businessName,
+            TaxNumber = _taxNumber
         };
 
-        var requestValidator = new DeleteBusinessCommandValidator();
+        var requestValidator = new AddBusinessCommandValidator();
 
         _validationErrors = await requestValidator.ValidateAsync(request);
 
         if (_validationErrors.IsValid)
             try
             {
-                var handler = new DeleteBusinessCommandHandler(userBusinessRepoMock.Object);
+                var handler = new AddBusinessCommandHandler(userBusinessRepoMock.Object);
                 await handler.Handle(request, CancellationToken.None);
                 _responseType = CommandResponseType.Successful;
             }
@@ -59,8 +67,8 @@ public class DeleteBusinessCommandStepDefinitions : TestBase
                         _commandErrors = validationException.Errors;
                         _responseType = CommandResponseType.BadRequest;
                         break;
-                    case NotFoundException notFoundException:
-                        _responseType = CommandResponseType.NotFound;
+                    case ConflictException conflictException:
+                        _responseType = CommandResponseType.Conflict;
                         break;
                     default:
                         _responseType = CommandResponseType.Error;
@@ -82,8 +90,8 @@ public class DeleteBusinessCommandStepDefinitions : TestBase
             case "BadRequest":
                 _responseType.Should().Be(CommandResponseType.BadRequest);
                 break;
-            case "NotFound":
-                _responseType.Should().Be(CommandResponseType.NotFound);
+            case "Conflict":
+                _responseType.Should().Be(CommandResponseType.Conflict);
                 break;
         }
     }
