+++ src/Subjects/Specs.Unit/Business/DeleteBusinessCommand.feature
