+++ src/Subjects/Specs.Unit/Business/DeleteBusinessCommand.feature.cs
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace Goodtocode.Subjects.Integration.Business.Commands
+namespace Goodtocode.Subjects.Unit.Business
 {
     using TechTalk.SpecFlow;
     using System;
@@ -20,25 +20,25 @@ namespace Goodtocode.Subjects.Integration.Business.Commands
     [System.CodeDom.Compiler.GeneratedCodeAttribute("TechTalk.SpecFlow", "3.9.0.0")]
     [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     [NUnit.Framework.TestFixtureAttribute()]
-    [NUnit.Framework.DescriptionAttribute("Update Business Command")]
-    [NUnit.Framework.CategoryAttribute("updateBusinessCommand")]
-    public partial class UpdateBusinessCommandFeature
+    [NUnit.Framework.DescriptionAttribute("Delete Business Command")]
+    [NUnit.Framework.CategoryAttribute("deleteBusinessCommand")]
+    public partial class DeleteBusinessCommandFeature
     {
         
         private TechTalk.SpecFlow.ITestRunner testRunner;
         
         private static string[] featureTags = new string[] {
-                "updateBusinessCommand"};
+                "deleteBusinessCommand"};
         
-#line 1 "UpdateBusinessCommand.feature"
+#line 1 "DeleteBusinessCommand.feature"
 #line hidden
         
         [NUnit.Framework.OneTimeSetUpAttribute()]
         public virtual void FeatureSetup()
         {
             testRunner = TechTalk.SpecFlow.TestRunnerManager.GetTestRunner();
-            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "Business/Commands", "Update Business Command", "As an customer service agent\r\nI can change business information\r\nAnd get the syst" +
-                    "em of record updated", ProgrammingLanguage.CSharp, featureTags);
+            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "Business", "Delete Business Command", "As an customer service agent\r\nI delete the business\r\nThe business is deleted from" +
+                    " the system of record", ProgrammingLanguage.CSharp, featureTags);
             testRunner.OnFeatureStart(featureInfo);
         }
         
@@ -77,15 +77,11 @@ namespace Goodtocode.Subjects.Integration.Business.Commands
         }
         
         [NUnit.Framework.TestAttribute()]
-        [NUnit.Framework.DescriptionAttribute("Update business")]
-        [NUnit.Framework.TestCaseAttribute("success TaxNumber Add", "Success", "", "d1604a05-f883-40f1-803b-8562b5674f1a", "businessName", "123-4567", null)]
-        [NUnit.Framework.TestCaseAttribute("success TaxNumber Update", "Success", "", "d1604a05-f883-40f1-803b-8562b5674f1a", "businessName", "123-4567", null)]
-        [NUnit.Framework.TestCaseAttribute("success TaxNumber Remove", "Success", "", "d1604a05-f883-40f1-803b-8562b5674f1a", "businessName", "", null)]
-        [NUnit.Framework.TestCaseAttribute("success Name Update", "Success", "", "d1604a05-f883-40f1-803b-8562b5674f1a", "businessName", "123-4567", null)]
-        [NUnit.Framework.TestCaseAttribute("bad request Invalid BusinessKey", "BadRequest", "BusinessKey", "11111", "businessName", "123-4567", null)]
-        [NUnit.Framework.TestCaseAttribute("bad request Invalid BusinessName", "BadRequest", "BusinessName", "d1604a05-f883-40f1-803b-8562b5674f1a", "", "123-4567", null)]
-        [NUnit.Framework.TestCaseAttribute("bad request", "BadRequest", "BusinessKey,BusinessName", "11111", "", "", null)]
-        public void UpdateBusiness(string def, string response, string responseErrors, string requestBusinessKey, string requestBusinessName, string requestTaxNumber, string[] exampleTags)
+        [NUnit.Framework.DescriptionAttribute("Delete business")]
+        [NUnit.Framework.TestCaseAttribute("success TaxNumber BusinessKey", "Success", "", "d1604a05-f883-40f1-803b-8562b5674f1a", null)]
+        [NUnit.Framework.TestCaseAttribute("bad request Empty BusinessKey", "BadRequest", "BusinessKey", "00000000-0000-0000-0000-000000000000", null)]
+        [NUnit.Framework.TestCaseAttribute("bad request non-guid BusinessKey", "BadRequest", "BusinessKey", "11111", null)]
+        public void DeleteBusiness(string def, string response, string responseErrors, string requestBusinessKey, string[] exampleTags)
         {
             string[] tagsOfScenario = exampleTags;
             System.Collections.Specialized.OrderedDictionary argumentsOfScenario = new System.Collections.Specialized.OrderedDictionary();
@@ -93,9 +89,7 @@ namespace Goodtocode.Subjects.Integration.Business.Commands
             argumentsOfScenario.Add("response", response);
             argumentsOfScenario.Add("responseErrors", responseErrors);
             argumentsOfScenario.Add("requestBusinessKey", requestBusinessKey);
-            argumentsOfScenario.Add("requestBusinessName", requestBusinessName);
-            argumentsOfScenario.Add("requestTaxNumber", requestTaxNumber);
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Update business", null, tagsOfScenario, argumentsOfScenario, featureTags);
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Delete business", null, tagsOfScenario, argumentsOfScenario, featureTags);
 #line 7
 this.ScenarioInitialize(scenarioInfo);
 #line hidden
@@ -113,18 +107,12 @@ this.ScenarioInitialize(scenarioInfo);
  testRunner.And(string.Format("I have a BusinessKey \"{0}\"", requestBusinessKey), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
 #line 10
- testRunner.And(string.Format("I have a BusinessName \"{0}\"", requestBusinessName), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
+ testRunner.When("I delete the business", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "When ");
 #line hidden
 #line 11
- testRunner.And(string.Format("I have a TaxNumber \"{0}\"", requestTaxNumber), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
-#line hidden
-#line 12
- testRunner.When("I update the business", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "When ");
-#line hidden
-#line 13
  testRunner.Then(string.Format("The response is \"{0}\"", response), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Then ");
 #line hidden
-#line 14
+#line 12
  testRunner.And(string.Format("If the response has validation issues I see the \"{0}\" in the response", responseErrors), ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
             }
