+++ src/Subjects/Specs.Unit/Specs.Unit.csproj
@@ -9,12 +9,6 @@
 		<Nullable>enable</Nullable>
 		<IsPackable>false</IsPackable>
 	</PropertyGroup>
-
-	<ItemGroup>
-		<SpecFlowObsoleteCodeBehindFiles Remove="Business\Commands\UpdateBusinessCommand - Copy.feature.cs" />
-		<SpecFlowObsoleteCodeBehindFiles Remove="Business\Queries\GetAllTicketedSessions.feature.cs" />
-		<SpecFlowObsoleteCodeBehindFiles Remove="Business\Queries\GetBusinessesByName - Copy.feature.cs" />
-	</ItemGroup>
 	<ItemGroup>
 		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="7.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="7.0.0" />
@@ -36,40 +30,40 @@
 		</None>
 	</ItemGroup>
 	<ItemGroup>
-		<Compile Update="Business\Commands\AddBusinessCommand.feature.cs">
+		<Compile Update="Business\AddBusinessCommand.feature.cs">
 		  <DependentUpon>AddBusinessCommand.feature</DependentUpon>
 		</Compile>
-		<Compile Update="Business\Commands\DeleteBusinessCommand.feature.cs">
+		<Compile Update="Business\DeleteBusinessCommand.feature.cs">
 		  <DependentUpon>DeleteBusinessCommand.feature</DependentUpon>
 		</Compile>
-		<Compile Update="Business\Commands\UpdateBusinessCommand.feature.cs">
+		<Compile Update="Business\UpdateBusinessCommand.feature.cs">
 			<DependentUpon>UpdateBusinessCommand.feature</DependentUpon>
 		</Compile>
-		<Compile Update="Business\Queries\GetBusinessesByKey.feature.cs">
+		<Compile Update="Business\GetBusinessesByKey.feature.cs">
 			<DependentUpon>GetBusinessesByKey.feature</DependentUpon>
 		</Compile>
-		<Compile Update="Business\Queries\GetBusinessesByName.feature.cs">
+		<Compile Update="Business\GetBusinessesByName.feature.cs">
 			<DependentUpon>GetBusinessesByName.feature</DependentUpon>
 		</Compile>
 	</ItemGroup>
 	<ItemGroup>
-		<SpecFlowFeatureFiles Update="Business\Commands\AddBusinessCommand.feature">
+		<SpecFlowFeatureFiles Update="Business\AddBusinessCommand.feature">
 		  <Visible>$(UsingMicrosoftNETSdk)</Visible>
 		  <CodeBehindFile>%(RelativeDir)%(Filename).feature$(DefaultLanguageSourceExtension)</CodeBehindFile>
 		</SpecFlowFeatureFiles>
-		<SpecFlowFeatureFiles Update="Business\Commands\DeleteBusinessCommand.feature">
+		<SpecFlowFeatureFiles Update="Business\DeleteBusinessCommand.feature">
 		  <Visible>$(UsingMicrosoftNETSdk)</Visible>
 		  <CodeBehindFile>%(RelativeDir)%(Filename).feature$(DefaultLanguageSourceExtension)</CodeBehindFile>
 		</SpecFlowFeatureFiles>
-		<SpecFlowFeatureFiles Update="Business\Commands\UpdateBusinessCommand.feature">
+		<SpecFlowFeatureFiles Update="Business\UpdateBusinessCommand.feature">
 			<Visible>$(UsingMicrosoftNETSdk)</Visible>
 			<CodeBehindFile>%(RelativeDir)%(Filename).feature$(DefaultLanguageSourceExtension)</CodeBehindFile>
 		</SpecFlowFeatureFiles>
-		<SpecFlowFeatureFiles Update="Business\Queries\GetBusinessesByKey.feature">
+		<SpecFlowFeatureFiles Update="Business\GetBusinessesByKey.feature">
 			<Visible>$(UsingMicrosoftNETSdk)</Visible>
 			<CodeBehindFile>%(RelativeDir)%(Filename).feature$(DefaultLanguageSourceExtension)</CodeBehindFile>
 		</SpecFlowFeatureFiles>
-		<SpecFlowFeatureFiles Update="Business\Queries\GetBusinessesByName.feature">
+		<SpecFlowFeatureFiles Update="Business\GetBusinessesByName.feature">
 			<Visible>$(UsingMicrosoftNETSdk)</Visible>
 			<CodeBehindFile>%(RelativeDir)%(Filename).feature$(DefaultLanguageSourceExtension)</CodeBehindFile>
 		</SpecFlowFeatureFiles>
